# RISC-V ASM: SHA-3