# Description

Add assembly implementations of SHA-3.
Use VSRL_VX instead of two VSRL_VI operations as immediate is only 5 bits.

# Testing

 --enable-riscv-asm --enable-sha3
--enable-riscv-asm=zbb --enable-sha3
--enable-riscv-asm=zv --enable-sha3
--enable-riscv-asm=zv,zvbb --enable-sha3
--enable-riscv-asm=zv,zvbb,zbb --enable-sha3

# Checklist

 - [ ] added tests
 - [ ] updated/added doxygen
 - [ ] updated appropriate READMEs
 - [ ] Updated manual and documentation