@dgarske 2024-08-30
```HiFive Unleased at 1.2GHz

./configure '--enable-riscv-asm' '--enable-sha3'

PR 7916
root@HiFiveU:~/wolfssl# ./wolfcrypt/benchmark/benchmark -sha3
------------------------------------------------------------------------------
 wolfSSL version 5.7.2
------------------------------------------------------------------------------
Math: 	Multi-Precision: Wolf(SP) word-size=64 bits=3072 sp_int.c
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
SHA3-224                    20 MiB took 1.130 seconds,   17.706 MiB/s
SHA3-256                    20 MiB took 1.196 seconds,   16.725 MiB/s
SHA3-384                    15 MiB took 1.134 seconds,   13.228 MiB/s
SHA3-512                    10 MiB took 1.064 seconds,    9.396 MiB/s
Benchmark complete

Master
root@HiFiveU:~/wolfssl# ./wolfcrypt/benchmark/benchmark -sha3
------------------------------------------------------------------------------
 wolfSSL version 5.7.2
------------------------------------------------------------------------------
Math: 	Multi-Precision: Wolf(SP) word-size=64 bits=3072 sp_int.c
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
SHA3-224                    15 MiB took 1.060 seconds,   14.154 MiB/s
SHA3-256                    15 MiB took 1.114 seconds,   13.468 MiB/s
SHA3-384                    15 MiB took 1.428 seconds,   10.506 MiB/s
SHA3-512                    10 MiB took 1.350 seconds,    7.408 MiB/s
Benchmark complete```